+++ src/components/CardRegisterForm/cardRegisterForm.module.css
@@ -1,0 +1,153 @@
+.form {
+  display: flex;
+  flex-direction: column;
+  row-gap: 16px;
+  margin-top: 20px;
+}
+
+.inputContainer {
+  display: flex;
+  align-items: center;
+  width: 318px;
+  height: 45px;
+  padding: 0 56px;
+  margin-top: 8px;
+  background: #ecebf1;
+  border: none;
+  border-radius: 7px;
+  text-align: center;
+}
+
+.form > label {
+  font-family: 'Roboto';
+  font-style: normal;
+  font-weight: 500;
+  font-size: 12px;
+  line-height: 14px;
+  color: #525252;
+}
+
+.inputContainer > input {
+  width: 25%;
+  background: transparent;
+  border: none;
+  text-align: center;
+  font-family: 'Roboto';
+  font-size: 16px;
+  font-weight: 600;
+  letter-spacing: 0.05em;
+}
+
+.inputContainer > input[type='password'] {
+  letter-spacing: -0.2em;
+}
+
+.expirationDate {
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  width: 137px;
+  height: 45px;
+  margin-top: 9px;
+  background: #ecebf1;
+  border: none;
+  border-radius: 7px;
+  text-align: center;
+  font-size: 13px;
+}
+
+.expirationDate > input {
+  width: 25%;
+  background: transparent;
+  border: none;
+  text-align: center;
+  font-family: 'Roboto';
+  font-size: 16px;
+  font-weight: 600;
+  letter-spacing: 0.05em;
+}
+
+.ownerName {
+  width: 318px;
+  height: 45px;
+  margin-top: 9px;
+  background: #ecebf1;
+  border: none;
+  border-radius: 7px;
+}
+
+.ownerName > input {
+  width: 100%;
+  height: 100%;
+  padding: 0 12px;
+  background: transparent;
+  border: none;
+  font-family: 'Roboto';
+  font-size: 16px;
+}
+
+.cvcContainer {
+  display: flex;
+  align-items: center;
+  column-gap: 13px;
+  margin-top: 9px;
+}
+
+.cvc {
+  width: 84px;
+  height: 45px;
+  background: #ecebf1;
+  border: none;
+  border-radius: 7px;
+}
+
+.cvc > input {
+  width: 100%;
+  height: 100%;
+  padding: 0 12px;
+  background: transparent;
+  border: none;
+  font-family: 'Roboto';
+  font-size: 16px;
+  text-align: center;
+}
+
+.cardPassword {
+  display: flex;
+  align-items: center;
+  column-gap: 7px;
+  width: 174px;
+  height: 45px;
+  margin-top: 9px;
+}
+
+.cardPassword > input {
+  width: 25%;
+  height: 100%;
+  padding: 0 12px;
+  background: #ecebf1;
+  border-radius: 7px;
+  border: none;
+  font-family: 'Roboto';
+  font-size: 16px;
+  text-align: center;
+}
+
+.cardPassword > p {
+  font-size: 24px;
+  font-weight: 900;
+}
+
+.submitButton {
+  width: 51px;
+  height: 34px;
+  margin-left: auto;
+  background: transparent;
+  border: none;
+  font-family: 'Roboto';
+  font-style: normal;
+  font-weight: 700;
+  font-size: 14px;
+  line-height: 16px;
+  cursor: pointer;
+}
